Host-app build works
