+++ angular-universal/host-app/server-bootstrap.ts
@@ -1,0 +1,56 @@
+import 'zone.js/dist/zone-node';
+
+import { ngExpressEngine } from '@nguniversal/express-engine';
+import * as express from 'express';
+import { join } from 'path';
+
+import { AppServerModule } from './src/main.server';
+import { APP_BASE_HREF } from '@angular/common';
+import { existsSync } from 'fs';
+
+// The Express app is exported so that it can be used by serverless Functions.
+export function app() {
+    const server = express();
+    const distFolder = join(process.cwd(), 'dist/host-app');
+    const indexHtml = existsSync(join(distFolder, 'index.original.html')) ? 'index.original.html' : 'index';
+    
+    // Our Universal express-engine (found @ https://github.com/angular/universal/tree/master/modules/express-engine)
+    server.engine('html', ngExpressEngine({
+        bootstrap: AppServerModule,
+    }));
+    
+    server.set('view engine', 'html');
+    server.set('views', distFolder);
+    
+    
+    server.get('/favicon.ico', (req, res) => {
+        res.status(404).send();
+    });
+    
+    // Example Express Rest API endpoints
+    // server.get('/api/**', (req, res) => { });
+    // Serve static files from /browser
+    server.get('*.*', express.static(distFolder, {
+        maxAge: '1y'
+    }));
+    
+    
+    // All regular routes use the Universal engine
+    server.get('*', (req, res) => {
+        res.render(indexHtml, { req, providers: [{ provide: APP_BASE_HREF, useValue: req.baseUrl }] });
+    });
+    
+    return server;
+}
+
+export function run() {
+    const port = process.env.PORT || 4000;
+    
+    // Start up the Node server
+    const server = app();
+    server.listen(port, () => {
+        console.log(`Node Express server listening on http://localhost:${port}`);
+    });
+}
+
+export * from './src/main.server';
