+++ angular-universal/host-app/server.js
@@ -1,0 +1,7 @@
+// declare const __non_webpack_require__: NodeRequire;
+
+const mainModule = __non_webpack_require__.main;
+const moduleFilename = mainModule && mainModule.filename || '';
+if (moduleFilename === __filename || moduleFilename.includes('iisnode')) {
+  import ('./server-bootstrap').then(x => x.run());
+}
