+++ angular-universal/host-app/webpack.config.js
@@ -1,60 +1,7 @@
-const {CleanWebpackPlugin} = require('clean-webpack-plugin');
-const {resolve} = require('path');
-const {AngularCompilerPlugin} = require('@ngtools/webpack');
-const ProgressPlugin = require('webpack/lib/ProgressPlugin');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-const ContainerReferencePlugin = require('webpack/lib/container/ContainerReferencePlugin');
+const host = require('./webpack/host-app');
+const hostSsr = require('./webpack/host-ssr');
 
-module.exports = (env = {}) => {
-    const buildFolder = `./dist/host`;
-
-    return {
-        entry: ['./src/polyfills.ts', './src/main.ts'],
-        mode: 'development',
-        output: {
-            path: resolve(__dirname, buildFolder),
-        },
-        plugins: [
-            // new CleanWebpackPlugin({
-            //     cleanOnceBeforeBuildPatterns: [buildFolder]
-            // }),
-            new ProgressPlugin(),
-            new HtmlWebpackPlugin(
-                {
-                    template: resolve(__dirname, './src/index.html'),
-                },
-            ),
-            new ContainerReferencePlugin({
-                remoteType: 'var',
-                remotes: {clientWeather: 'clientWeather'},
-                overrides: ['@angular/core', '@angular/common', '@angular/router']
-            }),
-            new AngularCompilerPlugin({
-                tsConfigPath: './tsconfig.app.json',
-                entryModule: './src/app/app.module#AppModule',
-                skipCodeGeneration: true,
-                directTemplateLoading: false,
-            }),
-        ],
-        module: {
-            rules: [
-                {
-                    test: /\.scss$/,
-                    use: ['raw-loader', 'sass-loader']
-                },
-                {
-                    test: /\.css$/,
-                    loader: 'raw-loader'
-                },
-                {
-                    test: /(?:\.ngfactory\.js|\.ngstyle\.js|\.ts)$/,
-                    loader: '@ngtools/webpack',
-                },
-                {
-                    test: /\.html$/,
-                    loader: 'raw-loader'
-                },
-            ],
-        },
-    };
-};
+module.exports = [
+    host,
+    hostSsr
+];
